+++ packages/UsageAnalysis/src/view-handler.ts
@@ -37,14 +37,17 @@ export class ViewHandler {
     const params = this.getSearchParameters();
     // [ErrorsView, FunctionsView, UsersView, DataView];
     const viewClasses: (typeof UaView)[] = [OverviewView, PackagesView, FunctionsView, EventsView, LogView, TestsView, ErrorsView, ReportsView];
+    const inits: Promise<void>[] = [];
     for (let i = 0; i < viewClasses.length; i++) {
       const currentView = new viewClasses[i](toolbox);
-      currentView.tryToinitViewers();
+      inits.push(currentView.tryToinitViewers());
       ViewHandler.UA.addView(currentView.name, () => currentView, false);
     }
+    await Promise.all(inits);
     const paramsHaveDate = params.has('date');
     const paramsHaveUsers = params.has('groups');
     const paramsHavePackages = params.has('packages');
+    const paramsHaveReport = params.has('report');
     if (paramsHaveDate || paramsHaveUsers || paramsHavePackages) {
       if (paramsHaveDate)
         toolbox.setDate(params.get('date')!);
@@ -143,9 +146,9 @@ export class ViewHandler {
     });
     ViewHandler.UA.name = ViewHandler.UAname;
     ViewHandler.UA.box = true;
-    const urlTab = window.location.pathname.match(/UsageAnalysis\/([a-zA-Z]+)/)?.[1];
-    ViewHandler.UA.path = APP_PREFIX + (urlTab ?? 'Overview');
-    if (urlTab) ViewHandler.changeTab(urlTab);
+    const urlTab = window.location.pathname.match(/UsageAnalysis\/UsageAnalysis\/([a-zA-Z]+)/)?.[1] ?? 'Overview';
+    ViewHandler.UA.path = `/${urlTab}`;
+    if (viewClasses.some((v) => v.name === `${urlTab}View`)) ViewHandler.changeTab(urlTab);
     grok.shell.addView(ViewHandler.UA);
   }